+++ ciudades/src/app/edicion-ciudad/edicion-ciudad.component.html
@@ -1,0 +1,20 @@
+<mat-card class="example-card">
+    <mat-card-title>Crear ciudad</mat-card-title>
+  
+    <mat-card-content class="contenedor-partes">
+      <div [formGroup]="formCity">
+        <div>
+          <mat-form-field>
+            <mat-label>Nombre</mat-label><input matInput placeholder="ciudad" formControlName="name">
+          </mat-form-field>
+        </div>
+        <div><mat-form-field>
+            <mat-label>Codigo Postal</mat-label><input matInput placeholder="codigo postal" formControlName="zipCode">
+          </mat-form-field>
+        </div>
+      </div>
+    </mat-card-content>
+    <mat-card-actions class="contenedor-partes"><button mat-stroked-button
+        (click)="guardarOActualizarCiudad()">GUARDAR</button></mat-card-actions>
+  </mat-card>
+  <hr>
